docker/guitar_lovers/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.8
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-filter==21.1
django-redis==5.1.0
djangorestframework==3.12.4
flower==1.0.0
gunicorn==20.1.0
humanize==3.13.1
idna==3.3
importlib-metadata==4.8.2
kombu==5.2.2
Markdown==3.3.5
oauthlib==3.1.1
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
psycopg2==2.9.1
pycparser==2.21
PyJWT==2.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tornado==6.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0